Abstract sd setup to a separate file.
